pyproject.toml
[tool.poetry]
name = "demo-app"
version = "0.1.0"
description = "FastAPI, Docker and Co"
authors = ["Sherevv <blog@sherevv.ru>"]

[tool.poetry.dependencies]
python = "^3.10"
alembic = "^1.7.7"
fastapi = "^0.78.0"
SQLAlchemy = "1.4.35"
sqlmodel = "^0.0.6"
asyncpg = "^0.25.0"
strawberry-graphql = {extras = ["fastapi"], version = "^0.113.0"}
uvicorn = {extras = ["standard"], version = "^0.17.6"}
celery = "^5.2.6"
flower = "^1.0.0"
redis = "^4.3.1"
gunicorn = "^20.1.0"

[tool.poetry.dev-dependencies]
pep8 = "^1.7.1"
pytest = "^7.1.2"
autopep8 = "^1.6.0"
pylint = "^2.13.9"
pytest-asyncio = "^0.18.3"
httpx = "^0.23.0"
asgi-lifespan = "^1.0.1"
trio = "^0.20.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
